Support --dry-run mode
When in dry-run mode, Cucumber does not seem to refresh its world after
each feature. This commit changes the code to create a new feature
instance when the currently running feature changes, instead of just
using ||= to create an instance on the "first" step of each feature.

Also, in dry-run mode, features/support/env.rb is not executed, as
documented in the Cucumber help. this commit changes the example project
to require the gem in the hooks.rb file instead.